worked on the login, still won't redirect
I tried a bunch of stuff including creating a function in a different screen and then calling it back once loggedin is true, but it won't work since loggedin is not defind in that screen. I tried to put the function in the email form screen which authinticates but it'll give me errors.